pyproject.toml
[project]
name = "polyconf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-benedict[toml,yaml]>=0.34.1",
    "voluptuous>=0.15.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/polyconf"]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
    "ptpython>=3.0.30",
]
